requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asteroid-filterbanks==0.4.0
asyncpg==0.31.0
attrs==25.4.0
audioread==3.1.0
bidict==0.23.1
billiard==4.2.4
blinker==1.9.0
brotli==1.2.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorlog==6.10.1
ConfigArgParse==1.7.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
deepgram-sdk==5.3.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
fastapi==0.128.0
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gevent==25.9.1
geventhttpclient==2.3.7
googleapis-common-protos==1.72.0
greenlet==3.3.1
groq==1.0.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.9
kombu==5.6.2
lazy_loader==0.4
librosa==0.11.0
lightning==2.6.0
lightning-utilities==0.15.2
limits==5.6.0
llvmlite==0.46.0
locust==2.43.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.6.1
noisereduce==3.0.3
numba==0.63.1
numpy==2.3.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.7.0
packaging==26.0
pandas==3.0.0
passlib==1.7.4
pgvector==0.4.2
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
pooch==1.8.2
primePy==1.3
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
psycopg2-binary==2.9.11
pyannote-audio==4.0.3
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-multipart==0.0.21
python-socketio==5.16.0
# pytorch-lightning==2.6.0
# pytorch-metric-learning==2.9.0
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
regex==2026.1.15
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.10.1
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.46
starlette==0.50.0
stripe==14.2.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
# torch-audiomentations==0.12.0
# torch_pitch_shift==1.2.5
# torchcodec==0.7.0
# torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.3.2
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.5
wrapt==2.0.1
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
zope.event==6.1
zope.interface==8.2
flower==2.0.1
minio==7.2.7
pytz==2024.1
bcrypt==4.1.2
